fix: kasync benchmarks
